Changes for v1.4.3

+++ database/migrations/2017_06_10_062632_add_foreign_keys_to_configurable_options_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 class AddForeignKeysToConfigurableOptionsTable extends Migration
 {
@@ -13,7 +14,13 @@ class AddForeignKeysToConfigurableOptionsTable extends Migration
     public function up()
     {
         Schema::table('configurable_options', function (Blueprint $table) {
-            $table->foreign('group_id')->references('id')->on('product_groups')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('configurable_options');
+            if (! array_key_exists('configurable_options_group_id_foreign', $indexesFound)) {
+                Schema::table('configurable_options', function (Blueprint $table) {
+                    $table->foreign('group_id')->references('id')->on('product_groups')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+                });
+            }
         });
     }
