+++ database/migrations/2017_06_10_062632_add_foreign_keys_to_order_invoice_relations_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 class AddForeignKeysToOrderInvoiceRelationsTable extends Migration
 {
@@ -13,8 +14,14 @@ class AddForeignKeysToOrderInvoiceRelationsTable extends Migration
     public function up()
     {
         Schema::table('order_invoice_relations', function (Blueprint $table) {
-            $table->foreign('invoice_id')->references('id')->on('invoices')->onUpdate('RESTRICT')->onDelete('RESTRICT');
-            $table->foreign('order_id')->references('id')->on('orders')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            $sm = Schema::getConnection()->getDoctrineSchemaManager();
+            $indexesFound = $sm->listTableIndexes('order_invoice_relations');
+            if (! array_key_exists('order_invoice_relations_invoice_id_foreign', $indexesFound)) {
+                $table->foreign('invoice_id')->references('id')->on('invoices')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            }
+            if (! array_key_exists('order_invoice_relations_order_id_foreign', $indexesFound)) {
+                $table->foreign('order_id')->references('id')->on('orders')->onUpdate('RESTRICT')->onDelete('RESTRICT');
+            }
         });
     }
